Add postgres ENV for db login
The databases could not be setup for this app, because they were
configured with postgres which requires a username and password
to run any of the database commands. Since writing these directly
into the source is not secure and should not be done, a separate
yaml file was created to store my database's username and password
as environmental variables (ENV) which could later be called upon
in the source code without revealing them to the public.