Add presence validation for title
The project model has no validations in place which meant that
completely empty projects could be created, which is not wanted.
A presence validation was added for the title of the project, so
that at least a title must be provided for a project to be created.
The description was not included, because I may want to add a project
without writing up the entire description at the time. The URL was
not included because some projects may not have a specific URL to
go to, at least not at the time of being added to this website.