Add controller for projects model
The application had a model setup for projects, but there was no
controller setup to guide the program where to go and what to do
in coordination and respect to these projects. So an empty Projects
controller was initiated and different actions for the model will
be added to the controller as needed throughout configuration.