Add index action
The controller for projects did not have any actions and since the
index actions, which is meant to display all of the projects on the
same page, is the most common action used, once was added. Currently
the action only assigns all projects in the database to an instance
variable projects. The rest of the necessary information and actions
will be added as I perform some TDD.