Change README content to match intent
The previous README only contained the standard garbage that is
initialized when the file is. I intend to use the website to display
some of my work and give an insight into who I am as a developer and
person, but the repository does not show what the website looks like.
So I added the general description of the site's purpose.