Create stylesheets folder and base CSS file
This Rails app did not have a folder for properly holding all
necessary stylesheets therefore one was created. This is where all
stylesheets will be held for the project so that they can easily
be tagged using Rails built in stylesheet link tag helper to quickly
and efficiently add appropriate stylesheets to any page.