Add basic site description to home page
The current home page had nothing on it and would therefore default
to the Rails app sample page, which is not very welcoming or
explanatory for visitors. So I added a simple description to the main
index page along with a note that the site is under construction so
that visitors do not automatically think that my site/code is broken.